+++ src/main/java/com/wanted/socialMediaIntegratedFeed/domain/post/PostHashtag.java
@@ -20,7 +20,7 @@ public class PostHashtag extends BaseTimeEntity {
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "content_id")
+    @JoinColumn(name = "post_id")
     private Post post;
 
     @ManyToOne(fetch = FetchType.LAZY)
