+++ src/main/java/com/wanted/socialMediaIntegratedFeed/domain/post/PostRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.wanted.socialMediaIntegratedFeed.domain.post;
 
-import com.wanted.socialMediaIntegratedFeed.web.content.dto.response.PostPaginationResponse;
+import com.wanted.socialMediaIntegratedFeed.web.post.dto.response.PostPaginationResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
